refactor: Embed enums into MemberEntity and remove external enum classes
