+++ entity/src/main/kotlin/com/mashup/dojo/MemberRepository.kt
@@ -1,6 +1,5 @@
 package com.mashup.dojo
 
-import com.mashup.dojo.member.MemberEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface MemberRepository : JpaRepository<MemberEntity, String>
