--- entity/src/main/kotlin/com/mashup/dojo/member/Gender.kt
@@ -1,6 +1,0 @@
-package com.mashup.dojo.member
-
-enum class Gender {
-    MALE,
-    FEMALE,
-}
