--- entity/src/main/kotlin/com/mashup/dojo/member/Platform.kt
@@ -1,9 +1,0 @@
-package com.mashup.dojo.member
-
-enum class Platform {
-    PRODUCT_DESIGN,
-    WEB,
-    IOS,
-    ANDROID,
-    SPRING,
-}
